+++ src/Client/Core/RELEASENOTES.md
@@ -1,1 +1,1 @@
-- Fix filter not being passed along in `PurgeAllInstancesAsync` (https://github.com/microsoft/durabletask-dotnet/pull/289)
+- Add new `IDurableTaskClientBuilder AddDurableTaskClient(IServiceCollection, string?)` API